Allow class methods to be queued on ActiveRecord objects.  Closes #41
